// front/src/app/finance/enums/fees-names.enum.ts
export enum FeesNames {
  oLevelApplicationFee = 'oLevelApplicationFee',
  aLevelApplicationFee = 'aLevelApplicationFee',
  deskFee = 'deskFee',
  oLevelTuitionDay = 'oLevelTuitionDay',
  aLevelTuitionDay = 'aLevelTuitionDay',
  oLevelTuitionBoarder = 'oLevelTuitionBoarder',
  aLevelTuitionBoarder = 'aLevelTuitionBoarder',
  oLevelScienceFee = 'oLevelScienceFee',
  alevelScienceFee = 'aLevelScienceFee',
  developmentFee = 'developmentFee',
  foodFee = 'foodFee',
  transportFee = 'transportFee',
  vacationTuitionDay = 'vacationTuitionDay',
  vacationTuitionBoarder = 'vacationTuitionBoarder',
  // Miscellaneous charges
  groomingFee = 'groomingFee',
  brokenFurnitureFee = 'brokenFurnitureFee',
  lostBooksFee = 'lostBooksFee',
  miscellaneousCharge = 'miscellaneousCharge',
}